requirements.txt
pandas>=2.0
numpy>=1.24
PyYAML>=6.0
